Set Up LWJGL and Slick2d
I set up LWJGL and Slick2d and got a window going.  Still need tests
done to see if it works on Mac, but it works on Windows well.  I set the
target fps to 60, but some people might want to choose their own, so
should we use a options.txt (or .ini which I've heard is commonly used
for options) for target FPS?