Flags!
Added Options.ini (current option is fullscreen), some state work-up,
and flags for each state.  There are two for Oregon since apparently
Oregon's flag has a separate front and back.  Some states will need to
be changed because they have the date that they became a state on it but
we'll need to change that to the date that they became a country.  This
will get complicated since the days can change.  It will have to be done
in-game.  Currently when you open the game it shows Texas' flag as proof
that flags can be loaded.