Fixed bug, made deciding nearly instantaneous
The instantaneousness is only on my computer, so it may be slower on a
weaker computer.  The probabilities are very inconsistent and it's
almost the same as just randomizing the 50 states and picking one, so
maybe some of the variables matter less and could be put away.  It will
take some figuring out however.